refactor(format): split service.go into parse.go and serialize.go
